pyproject.toml
[tool.poetry]
name = "autograder"
version = "0.1.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = "^0.23.0"
slack-sdk = "^3.21.3"
python-dotenv = "^1.0.0"
psycopg = { version = "^3.1.9", extras = ["binary", "pool"] }


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.280"
pytest = "^7.4.0"
httpx = "^0.24.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
